Additional update: - Refactor to allow delegate to be placed in its own package - Add extra guard to prevent addMethodToDenyList from being called within OgnlRuntime.
